% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_series_function.R
\name{align_series}
\alias{align_series}
\title{Align samples that statistically crossdate}
\usage{
align_series(the_data, cross_dates, sel_target)
}
\arguments{
\item{the_data}{A dataframe containing the individual seeries to be aligned.}

\item{cross_dates}{A dataframe containing the results of the crossdating (generated using lead_lag_analysis()[1]).
These data should be filtered, using filter_crossdates(), so this dataframe only contains significant ccrossdates
to the selected target sample.}

\item{sel_target}{A sample ID (column name) entered as a character string. This must be the same as the target sample used in the filter_crossdates() function.}
}
\description{
Align samples that have been crossdated using a target sample.
}
\keyword{Crossdate}
