% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_to_chron_function.R
\name{align_to_chron}
\alias{align_to_chron}
\title{Align crossdated series to series contained in a chronology}
\usage{
align_to_chron(the.data, chrono)
}
\arguments{
\item{the.data}{The dataframe containing the aligned data with the arithemtic mean chronology (in the second column)
and the undated series that hve been aligned in time with the mean chronology.}

\item{chrono}{A dataframe containing the individual series used to construct the arithmetic mean chronology}
}
\description{
This function replaces the arithmetic mean chronology, which has been aligned with undated series,
with the individual series used to generate the arithmetic mean chronology.
}
\keyword{crossdating}
