% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_undaed_load_function.R
\name{align_undated_load}
\alias{align_undated_load}
\title{Align samples in time for loading}
\usage{
align_undated_load(undated_df_data, loading_df_data)
}
\arguments{
\item{undated_df_data}{A dataftrame containing the previously loaded data. Should be at least two columns, first column the year values and rest growth data}

\item{loading_df_data}{A dataftrame containing the new data to load. Should be at least two columns, first column the year values and rest growth data}
}
\description{
Align samples in time for the loading processes
}
\examples{

undated_df_data  <- data.frame(years = 10:50, sample_a = runif(41))
loading_df_data <- data.frame(years = 25:75, sample_b = runif(51))
align_undated_load(undated_df_data, loading_df_data)
}
\keyword{aligning}
\keyword{dataframes}
\keyword{in}
\keyword{time}
