% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_correl.R
\name{auto_correl}
\alias{auto_correl}
\title{Calculate the autocorrelation contained in a series}
\usage{
auto_correl(the_data)
}
\arguments{
\item{the_data}{A dataframe containing series to be analysed.}
}
\description{
Calculate the autocorrelation (1-10th orders) contained in a series.
}
\keyword{autocorrelation}
