% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ring_measurer_fun.R
\name{check_load_ringmeasurer_data}
\alias{check_load_ringmeasurer_data}
\title{Load CSV data either from Ring Measurer App or from abother source}
\usage{
check_load_ringmeasurer_data(the_data, avg_series = TRUE)
}
\arguments{
\item{the_data}{A dataframe containing ring width measurements}

\item{file}{A string holding the path to the data. used to get the sample ID that should be the file name}
}
\description{
This function creates a dataframe containing ring width data.
}
\examples{
undated_path <- system.file("extdata", "undated_example.csv", package="ringdater")
the_data <- read.csv(undated_path, header = TRUE, stringsAsFactors = FALSE, check.names = FALSE)
check_load_ringmeasurer_data(undated_path)
}
\keyword{RingdateR}
