% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_pal_function.R
\name{col_pal}
\alias{col_pal}
\title{Set colours for RingdateR plots}
\usage{
col_pal(colour_scale = 1)
}
\arguments{
\item{colour_scale}{A numeric from 1 to 4. 1 = blue to white to red;
2 = white to red; 3 = white to blue; 4 = white to black.}
}
\description{
Sets the colour pallate for the heatmaps in RingdateR
}
\details{
1 = blue to white to red;
2 = white to red;
3 = white to blue;
4 = white to black.
}
\examples{
col_pal(1)
}
\keyword{Colours}
