% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_fun.R
\name{comb.NA}
\alias{comb.NA}
\title{replacement for Cbind.fill}
\usage{
comb.NA(..., fill = NA)
}
\arguments{
\item{...}{Either a dataframe and a vector or two data frames to be combined.}

\item{fill}{Either NA or a character string to be used to extend columns}
}
\description{
Combines vector with a dataframe of a different length
}
\examples{
df1<-data.frame(a=1:10, b=2:11)
df2<-data.frame(c=5:10, d=6:11)
comb.NA(df1,df2)
}
\keyword{autocorrelation}
