% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_RM_files_function.R
\name{combine_RM_files}
\alias{combine_RM_files}
\title{Loada bunch of Ring MEasurer measurement files in to a single CSV file}
\usage{
combine_RM_files(path)
}
\arguments{
\item{path}{A string containing the path for a directory containing measurement files - these can be in subdirectoriers}
}
\description{
This function creates a dataframe containing ring width data.
}
\examples{
undated_path <- system.file("extdata", "undated_example.csv", package="ringdater")
the_data <- read.csv(undated_path, header = TRUE, stringsAsFactors = FALSE, check.names = FALSE)
check_load_ringmeasurer_data(undated_path)

}
\keyword{Measurer}
\keyword{Ring}
\keyword{RingdateR}
