% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correl_replace.R
\name{correl_replace}
\alias{correl_replace}
\title{Evaluate correlations with replacement against a mean chronology}
\usage{
correl_replace(the.data)
}
\arguments{
\item{the.data}{A dataframe containing at least three series and a year column}
}
\description{
Calculate the correlation of each sample, used to construct an arithmetic
mean chronology, against the arithmetic mean chronology constructed excluding
the selected sample
}
\keyword{correlations}
