% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_date_plot.R
\name{dated_line_plot}
\alias{dated_line_plot}
\title{RingdateR shiny server file}
\usage{
dated_line_plot(the.data)
}
\arguments{
\item{the.data}{a dataframe to be organised so it can be plotted}
}
\description{
This function creates the shiny user interface
}
\keyword{GUI}
