% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrending_plot_fun.R
\name{detrending.plot.fun}
\alias{detrending.plot.fun}
\title{Generate a plot of the dtrended data}
\usage{
detrending.plot.fun(
  undet.data,
  first_series,
  detrending_select = 3,
  splinewindow = 21,
  font_size = 12,
  axis_line_width = 1,
  plot_line = 1
)
}
\arguments{
\item{undet.data}{A data.frame containing at least two columns. First column years/ring number second column ring widths. Can be morethan two columns of ring width data.}

\item{first_series}{A column name}

\item{detrending_select}{A numeric integer to define which detrending method to use. 1 = Do nothing, 2 = Z-scores, 3 = spline detrending, 4 = Mod. negative exponentia, 5 = Friedman, 6 = ModHugershoff, 7 = First difference. Spline is slected by default.}

\item{splinewindow}{A numeric integer to define the length of the spline to be used in splne detrending (if option 3 selected).}

\item{font_size}{A numeric integer to define the size of axis labels}

\item{axis_line_width}{A numeric integer to define the thickness ofthe axis lines}

\item{plot_line}{A numeric integer to define the thickness of the plotted lines}
}
\description{
This function produces a plot of the raw and detrended ring width data as well as the corresponding autocorrelation.
}
\examples{
path <- system.file("extdata", "undated_example.csv", package="ringdater")
the_data <- load_undated(path)
series <- colnames(the_data)[2]
detrending.plot.fun(undet.data = the_data,
                   first_series = series,
                   detrending_select = 3,
                   splinewindow = 21)
}
\keyword{image}
\keyword{pos,}
\keyword{pro}
