% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_crossdates_func.R
\name{filter_crossdates}
\alias{filter_crossdates}
\title{RingdateR shiny server file}
\usage{
filter_crossdates(
  the_data,
  r_val = 0.5,
  p_val = 0.05,
  overlap = 50,
  target = NULL
)
}
\arguments{
\item{the_data}{A dataframe created by lead_lag_analysis() function (lead_lag_analysis()[1]). Must be a dataframe.}

\item{r_val}{A numeric value (>=0 and <=1). Correlation coefficient.}

\item{p_val}{A numeric value (>=0 and <=1). Probablility value.}

\item{overlap}{An integer >0. This value represents the period of overlap between samples.}

\item{target}{a column name to be used as the target sample}
}
\description{
This function filters the results table created by A dataframe created by lead_lag_analysis() function (lead_lag_analysis()[1])
}
\keyword{statistics}
\keyword{tests}
