% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_analysis_function.R
\name{heatmap_analysis}
\alias{heatmap_analysis}
\title{running lead_lag analysis heat maps}
\usage{
heatmap_analysis(
  the_data,
  s1,
  s2,
  neg_lag = -20,
  pos_lag = 20,
  win = 21,
  center = 0,
  complete = TRUE,
  sel_col_pal = 1,
  font_size = 12,
  axis_line_width = 1,
  plot_line = 1,
  leg_size = 1
)
}
\arguments{
\item{the_data}{A dataframe containing the timeseries. First column should contain dates.}

\item{s1}{the name of a sample as a character string - must match a column name exactly}

\item{s2}{the name of a sample  as a character string - must match a column name exactly}

\item{neg_lag}{An integer}

\item{pos_lag}{An integer}

\item{win}{The window with which the running correlations are calculated over}

\item{center}{A numeric integer used to center the lead lag limits around. e.g. the pos lag is added to this value.}

\item{complete}{A boolean to classify whther to run the lead-lag over the maximum range of leads and lags}

\item{sel_col_pal}{The colour pallate to use (numeric)}

\item{font_size}{The size of the font (numeric)}

\item{axis_line_width}{The width of the axis (numeric)}

\item{plot_line}{The width of the line (numeric)}

\item{leg_size}{A numeric integer to define the size of the colour bar legend.}
}
\description{
This function rperforms the full running lead-lag analysis and produces a plot of the results
}
\examples{
undated_path <- system.file("extdata", "undated_example.csv", package="ringdater")
undated_data <- load_undated(undated_path)
undated_data <- name_check(undated_data)
undated      <- normalise(the.data = undated_data, detrending_select = 3, splinewindow = 21)
heatmap_analysis(the_data = undated, s1 = colnames(undated)[2], s2 = colnames(undated)[3],
                 neg_lag = -20, pos_lag = 20, win = 21, complete = FALSE)
}
\keyword{pairwise_lead_lag}
