% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_undated_chron.R
\name{ld_undated_chron}
\alias{ld_undated_chron}
\title{Load multiple undated chronologies into a single dataframe}
\usage{
ld_undated_chron(
  files,
  series_names = NULL,
  pair_detrend = FALSE,
  detrending_select = 3,
  splinewindow = 21,
  powerT = FALSE,
  shiny = FALSE
)
}
\arguments{
\item{files}{A list of file paths}

\item{series_names}{A list of file names. If not using the shiny app,  series names are automatically generated in the function so do not need to be parsed.}

\item{pair_detrend}{A boolean to dictate whether detrending should be applied or not. Set to false by default.}

\item{detrending_select}{Select the method for detrending. 1 = Do nothing, 2 = Z-scores, 3 = spline detrending, 4 = Mod. negative exponentia, 5 = Friedman, 6 = ModHugershoff, 7 = First difference. Spline is slected by default.}

\item{splinewindow}{An integer to define the window to be used if spline detrending is selected. Set to 21 years by default.}

\item{powerT}{A boolean to dictate whether or not to power transform the data.}

\item{shiny}{A boolean to let the function know if it is being run in the shiny app or in the console.}
}
\description{
Create a dataframe containing the arithmetic mean chronologies from multiple files
containing either two column chronologies or files containing multiple series to be
compiled as a chronology.
}
\keyword{RingdateR}
\keyword{load}
