% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pairwise_lead_lag.R
\name{lead_lag_analysis}
\alias{lead_lag_analysis}
\title{lead-lag analysis}
\usage{
lead_lag_analysis(
  the_data,
  mode = 1,
  neg_lag = -20,
  pos_lag = 20,
  complete = TRUE,
  shiny = TRUE
)
}
\arguments{
\item{the_data}{A dataframe containing the timeseries. First column should contain dates.}

\item{mode}{either 1 or 2. 1 = Pairwise analysis mode, 2 = Chronology analysis mode.}

\item{neg_lag}{An integer}

\item{pos_lag}{An integer}

\item{complete}{A boolean to classify whther to run the lead-lag over the maximum range of leads and lags}

\item{shiny}{defines whether the function is being called in the shiny app or not.}
}
\description{
This function runs lead-lag correlation analysis between multiple timeseries
}
\keyword{pairwise_lead_lag}
