% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_lag_bar_func.R
\name{lead_lag_bar}
\alias{lead_lag_bar}
\title{Produce a bar chart of leaad-lag correlations and show the best three matches.}
\usage{
lead_lag_bar(the_data, sample_1, sample_2)
}
\arguments{
\item{the_data}{A dataframe created by lead_lag_analysis() function (lead_lag_analysis()[2]).}

\item{sample_1}{A character string containing a sample ID. Must match the original sample IDs perfectly.}

\item{sample_2}{A character string containing a sample ID. Must match the original sample IDs perfectly.}
}
\description{
This function creates are bar chart ofthe lead-lag analyses from RingdateR
}
\keyword{bar}
\keyword{chart}
\keyword{lead-lag}
