% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot_fun.R
\name{line_plot}
\alias{line_plot}
\title{Line plot of two samples adjusted for lag}
\usage{
line_plot(the_data, series_1_nm, series_2_nm, lag = 0)
}
\arguments{
\item{the_data}{A data.frame containing the series to be plotted. The first column should contain years/ring numbers}

\item{series_1_nm}{A character string containing a column name in the the_data to be plotted.}

\item{series_2_nm}{A character string containing a column name in the the_data to be plotted. This series will be lagged.}

\item{lag}{A numeric integerto lag series series_2_nm by.}
}
\description{
PLot two timeseries and adjust the lag of the second series
}
\keyword{analysis}
\keyword{lead-lag}
