% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chron_fun.R
\name{load_chron}
\alias{load_chron}
\title{Load single chronology as a data frame}
\usage{
load_chron(file)
}
\arguments{
\item{file}{A  files path as a character string}
}
\description{
This function creates a dataframe containing the data from a dated chronology.
}
\keyword{RingdateR}
