% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_tabs_function.R
\name{load_data_tabs}
\alias{load_data_tabs}
\title{Produce small summary table of a dataframe}
\usage{
load_data_tabs(the_data)
}
\arguments{
\item{the_data}{A dataframe containing ring width series to be summarised.}
}
\description{
This function creates the loaded data summary tables.
}
\examples{
undated_path <- system.file("extdata", "undated_example.csv", package="ringdater")
undated_data <- load_undated(undated_path)
load_data_tabs(undated_data)
}
\keyword{summary}
