% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lps_function.R
\name{load_lps}
\alias{load_lps}
\title{Load Image Pro line profile series}
\usage{
load_lps(series, path)
}
\arguments{
\item{series}{The sample ID as a character string.}

\item{path}{The file path as a character string.}
}
\description{
This function loads lps measurement files created by Image pro as a data.frame
}
