% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pos_function.R
\name{load_pos}
\alias{load_pos}
\title{Load a pos file as a dataframe}
\usage{
load_pos(file_path)
}
\arguments{
\item{file_path}{A character string containing the file path to be loaded.}
}
\description{
This function reads pos formated files created by Image Pro and outputs a data.frame
}
\keyword{image}
\keyword{pos,}
\keyword{pro}
