% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_undated_function.R
\name{load_undated}
\alias{load_undated}
\title{Load multiple undated series into a single data frame}
\usage{
load_undated(files, col1 = "ring", series_names = NULL, shiny = FALSE)
}
\arguments{
\item{files}{A list of set of files pathsas a character strings}

\item{col1}{A character string to label the first column (e.g. Year or Ring)}

\item{series_names}{The file names, only used in the shiny server function.}

\item{shiny}{a logical to set series names. If running in cosole can be ignored.}
}
\description{
This function creates a dataframe containing the data from multiple files loaded at one.
}
\examples{
undated_path <- system.file("extdata", "undated_example.csv", package="ringdater")
load_undated(undated_path)
}
\keyword{RingdateR}
