% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRWL_functions.R
\name{locateID}
\alias{locateID}
\title{locateID}
\usage{
locateID(fName)
}
\arguments{
\item{fName}{a string containing a file path}
}
\description{
find sample IDs within RWL files
}
\keyword{data}
\keyword{load}
\keyword{rwl}
