% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_check_fun.R
\name{name_check}
\alias{name_check}
\title{Check the names of samples laoded into RingdateR}
\usage{
name_check(the_data)
}
\arguments{
\item{the_data}{A dataframe to check the colnames.}
}
\description{
This function checksthe column names of samples loaded into RingdateR
}
