% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onto_align_dated_function.R
\name{onto_align_dated}
\alias{onto_align_dated}
\title{Ontogenetically align series that are currently date aligned}
\usage{
onto_align_dated(df)
}
\arguments{
\item{df}{A dataframe containing the date aligned series to be aligned.}
}
\description{
Align samples that have been crossdated using a target sample.
}
\keyword{Crossdate,}
\keyword{align}
\keyword{ontogenetic,}
