% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_data_check_function.R
\name{pairwise_data_check}
\alias{pairwise_data_check}
\title{pairwise data check}
\usage{
pairwise_data_check(the_data)
}
\arguments{
\item{the_data}{A data frame containing the undated series}
}
\description{
Check if the are any problems with the loaded undated series
}
\keyword{RingdateR}
\keyword{load}
