% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_series_function.R
\name{plot_all_series}
\alias{plot_all_series}
\title{Plot all data in dataframe along with the arithmetic mean chronology}
\usage{
plot_all_series(aligned_data, text.size = 12, plot.line = 0.5, line.width = 1)
}
\arguments{
\item{aligned_data}{A data frame containing multiple ring width series aligned in time. First column should be years/ring count}

\item{text.size}{A numeric to set the axis font size}

\item{plot.line}{A numeric value to set the width of the lines being plotted}

\item{line.width}{A numeric to set the width of the line}
}
\description{
Generate a plot of all the aligned data along with the arithmetic mean chronology
}
\examples{
chron_path  <- system.file("extdata", "dated_example_excel.xlsx", package="ringdater")
chron_data  <- load_chron(chron_path)
chron_data  <- name_check(chron_data)
chrono      <- normalise(the.data = chron_data, detrending_select = 3, splinewindow = 21)

plot_all_series(chrono)
}
\keyword{plot}
