% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_sing_hm_function.R
\name{plotting_sing_hm}
\alias{plotting_sing_hm}
\title{Plot running lead_lag analysis}
\usage{
plotting_sing_hm(
  plot.data,
  the_data,
  s1,
  s2,
  font_size = 12,
  axis_line_width = 0.5,
  plot_line = 0.5,
  neg_lag = -20,
  pos_lag = 20,
  sel_col_pal = 1,
  leg_size = 1
)
}
\arguments{
\item{plot.data}{a three column data frame containing the results from the running_lead_lag function}

\item{the_data}{A dataframe containing the timeseries. First column should contain dates.}

\item{s1}{the name of a sample - must match a column name exactly}

\item{s2}{the name of a sample - must match a column name exactly}

\item{font_size}{The size of the font (numeric)}

\item{axis_line_width}{The width of the axis (numeric)}

\item{plot_line}{The width of the line (numeric)}

\item{neg_lag}{An integer}

\item{pos_lag}{An integer}

\item{sel_col_pal}{The colour pallet to use (numeric)}

\item{leg_size}{a numeric value to define the length of the colour scale legend}
}
\description{
This function plots the running lead_lag analysis
}
\keyword{pairwise_lead_lag}
