% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_check_function.R
\name{prob_check}
\alias{prob_check}
\title{Align samples that statistically crossdate}
\usage{
prob_check(new.chrono, wind = 20)
}
\arguments{
\item{new.chrono}{A dataframe containing the aligned series to be checked.}

\item{wind}{An integer to define the window length that the running analyses are calculated over}
}
\description{
Align samples that have been crossdated using a target sample.
}
\keyword{problems}
