% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_bar_EPS.R
\name{R_bar_EPS}
\alias{R_bar_EPS}
\title{Calculate the Rbar and EPS of a chronology}
\usage{
R_bar_EPS(the.data, window = 25)
}
\arguments{
\item{the.data}{A data.frame holding individual series in the aligned chronology}

\item{window}{A numeric integer used to calculate the running Rbar and EPS values over.}
}
\description{
This function uses dplR to calculate the Rbar and EPS of a chronology
}
\keyword{EPS}
\keyword{Rbar}
\keyword{chronology}
