% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRWL_functions.R
\name{readRWL}
\alias{readRWL}
\title{readRWL}
\usage{
readRWL(fName)
}
\arguments{
\item{fName}{a string containing a file path}
}
\description{
robust load data from rwl files
}
\keyword{data}
\keyword{load}
\keyword{rwl}
