% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRWL_functions.R
\name{readWOheader}
\alias{readWOheader}
\title{readWOheader}
\usage{
readWOheader(fName, startPT)
}
\arguments{
\item{fName}{a string containing a file path}

\item{startPT}{a string containing the start of a sample ID}
}
\description{
find and remove header from rwl files
}
\keyword{data}
\keyword{load}
\keyword{rwl}
