% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RingServer_function.R
\name{RingServer}
\alias{RingServer}
\title{This function is the backend code for the RingdateR shiny app.}
\usage{
RingServer(input, output, session)
}
\arguments{
\item{input}{Inputs are generated by the UI file}

\item{output}{Outputs are generated by the UI file}

\item{session}{Session is automatically created upon starting the shiny app}
}
\description{
This function is the backend code for the RingdateR shiny app.
}
\keyword{RingdateR}
\keyword{Shiny}
