% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollcor_function.R
\name{rollcor}
\alias{rollcor}
\title{rollcor}
\usage{
rollcor(x, y, width, show = TRUE, ...)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{width}{The width of the sliding window, which must be odd number.}

\item{show}{If TRUE, the result will be plotted.}

\item{...}{argument to pass into cor function}
}
\description{
performs running correlation analysis
}
