% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ringdater_func.R
\name{run_ringdater}
\alias{run_ringdater}
\title{Run RingdateR as a shiny app}
\usage{
run_ringdater()
}
\description{
This function runs RingdateR as a shiny app
}
\keyword{Shiny}
\keyword{app}
\keyword{crossdating}
