% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_lead_lag_analysis.R
\name{running_lead_lag}
\alias{running_lead_lag}
\title{running lead_lag analysis}
\usage{
running_lead_lag(
  the_data,
  s1,
  s2,
  neg_lag = -20,
  pos_lag = 20,
  win = 21,
  complete = TRUE
)
}
\arguments{
\item{the_data}{A dataframe containing the timeseries. First column should contain dates.}

\item{s1}{the name of a sample - must match a column name exactly}

\item{s2}{the name of a sample - must match a column name exactly}

\item{neg_lag}{An integer}

\item{pos_lag}{An integer}

\item{win}{The window with which the running correlations are calculated over}

\item{complete}{A boolean to classify whther to run the lead-lag over the maximum range of leads and lags}
}
\description{
This function rperforms running lead-lag correlation analysis between two timeseries
}
\keyword{pairwise_lead_lag}
