% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_fun.R
\name{vertLen}
\alias{vertLen}
\title{Calculate the length of a dataframe}
\usage{
vertLen(dfIn)
}
\arguments{
\item{dfIn}{A dataframe that is being combined}
}
\description{
Calculatesthe length of a dataframe
}
\keyword{cbind.fill}
