% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiten_function.R
\name{whitenSeries}
\alias{whitenSeries}
\title{prewhiten data}
\usage{
whitenSeries(series1)
}
\arguments{
\item{series1}{The file names, only used in the shiny server function.}
}
\description{
This function prewhitens series using COFECHA methodology
}
\keyword{RingdateR}
