% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_scale_bar_function.R
\name{x.scale.bar}
\alias{x.scale.bar}
\title{Generate a set of breaks for an x-axis}
\usage{
x.scale.bar(x.min, x.max)
}
\arguments{
\item{x.min}{a numeric integer}

\item{x.max}{a numeric integer}
}
\description{
This function sets the breaks for the x-axis in the RingdateR plots
}
\keyword{scale_x_Axis}
