% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_scale_bar_function.R
\name{y.scale.bar}
\alias{y.scale.bar}
\title{Generate a set of breaks for a y-axis}
\usage{
y.scale.bar(y.min, y.max)
}
\arguments{
\item{y.min}{a numeric integer}

\item{y.max}{a numeric integer#'}
}
\description{
his function sets the breaks for the y-axis in the RingdateR plots
}
\keyword{scale_x_Axis}
