% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_series_function.R
\name{align_series}
\alias{align_series}
\title{Align samples that statistically crossdate}
\usage{
align_series(the_data, cross_dates, sel_target)
}
\arguments{
\item{the_data}{A dataframe containing the individual seeries to be aligned.}

\item{cross_dates}{A dataframe containing the results of the crossdating (generated using lead_lag_analysis()[1]).
These data should be filtered, using filter_crossdates(), so this dataframe only contains significant ccrossdates
to the selected target sample.}

\item{sel_target}{A sample ID (column name) entered as a character string. This must be the same as the target sample used in the filter_crossdates() function.}
}
\description{
Align samples that have been crossdated using a target sample.
}
\examples{
undated_path <- system.file("extdata", "undated_example.csv", package="ringdater")
undated_data <- load_undated(undated_path)
undated      <- normalise(the.data = undated_data, detrending_select = 3, splinewindow = 21)

chron_path  <- system.file("extdata", "dated_example_excel.xlsx", package="ringdater")
chron_data  <- load_chron(chron_path)
chrono      <- normalise(the.data = chron_data, detrending_select = 3, splinewindow = 21)
chrono      <- data.frame(chrono[,1], rowMeans(chrono[,-1], na.rm = TRUE))
colnames(chrono)<-c("year", "mean_chronology")

chron_n_series <- comb.NA(chrono, undated[,-1], fill = NA)

exmp_analysis_1 <- system.file("extdata", "chron_comp_1.csv", package="ringdater")
chron_comp_1 <- read.csv(exmp_analysis_1, stringsAsFactors = FALSE, header = TRUE)

exmp_analysis_2 <- system.file("extdata", "chron_comp_2.csv", package="ringdater")
chron_comp_2 <- read.csv(exmp_analysis_2, stringsAsFactors = FALSE, header = TRUE)

chron_comp <- list(chron_comp_1,chron_comp_2)

filtered_data <- filter_crossdates(the_data = as.data.frame(chron_comp[1]),
                                   r_val = 0.4,
                                   p_val = 0.05,
                                   overlap = 30,
                                   target = "mean_chronology")

aligned_data <- align_series(the_data = chron_n_series,
                             cross_dates = filtered_data,
                             sel_target = "mean_chronology")
}
\keyword{Crossdate}
