% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dated_line_plot_function.R
\name{dated_line_plot}
\alias{dated_line_plot}
\title{Produces sample distribution plot data}
\usage{
dated_line_plot(the.data)
}
\arguments{
\item{the.data}{a dataframe to be organised so it can be plotted}
}
\description{
Organise data for plotting the distribution plot
}
\examples{
chron_path  <- system.file("extdata", "dated_example_excel.xlsx", package="ringdater")
chron_data  <- load_chron(chron_path)
plot_data <- dated_line_plot(chron_data)
ggplot2::ggplot(data = plot_data, ggplot2::aes(x=dates, y=samp.val, group = name.val))+
ggplot2::geom_line() +ggplot2::theme_minimal()
}
\keyword{data}
\keyword{organisation}
