% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_lag_bar_func.R
\name{lead_lag_bar}
\alias{lead_lag_bar}
\title{Produce a bar chart of leaad-lag correlations and show the best three matches.}
\usage{
lead_lag_bar(the_data, sample_1, sample_2)
}
\arguments{
\item{the_data}{A dataframe created by lead_lag_analysis() function (lead_lag_analysis()[2]).}

\item{sample_1}{A character string containing a sample ID. Must match the original sample IDs perfectly.}

\item{sample_2}{A character string containing a sample ID. Must match the original sample IDs perfectly.}
}
\description{
This function creates are bar chart ofthe lead-lag analyses from RingdateR
}
\examples{
# undated_path <- system.file("extdata", "undated_example.csv", package="ringdater")
# undated_data <- load_undated(undated_path)
# undated      <- normalise(the.data = undated_data, detrending_select = 3, splinewindow = 21)
#
# chron_path  <- system.file("extdata", "dated_example_excel.xlsx", package="ringdater")
# chron_data  <- load_chron(chron_path)
# chrono      <- normalise(the.data = chron_data, detrending_select = 3, splinewindow = 21)
# chrono      <- data.frame(chrono[,1], rowMeans(chrono[,-1], na.rm = TRUE))
# colnames(chrono)<-c("year", "mean_chronology")
# chron_n_series <- comb.NA(chrono, undated[,-1], fill = NA)
# chron_n_series[1:6,1:6]
#
# chron_comp <- lead_lag_analysis(the_data = chron_n_series,
#                                 mode = 2,
#                                 complete = TRUE,
#                                 shiny = FALSE)
# For speed, the results of the analyses above are imported as two CSV files and put
# together into a list (chron_comp).
exmp_analysis_1 <- system.file("extdata", "chron_comp_1.csv", package="ringdater")
chron_comp_1 <- read.csv(exmp_analysis_1, stringsAsFactors = FALSE, header = TRUE)

exmp_analysis_2 <- system.file("extdata", "chron_comp_2.csv", package="ringdater")
chron_comp_2 <- read.csv(exmp_analysis_2, stringsAsFactors = FALSE, header = TRUE)

chron_comp<-list(chron_comp_1,chron_comp_2)

lead_lag_bar(the_data = as.data.frame(chron_comp[2]),
             sample_1 = "mean_chronology",
             sample_2 = "sample_a")
}
\keyword{bar}
\keyword{chart}
\keyword{lead-lag}
