% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chron_function.R
\name{load_chron}
\alias{load_chron}
\title{Load single chronology as a data frame}
\usage{
load_chron(file)
}
\arguments{
\item{file}{A  files path as a character string}
}
\description{
This function creates a dataframe containing the data from a dated chronology.
}
\examples{
chron_path <- system.file("extdata", "dated_example_excel.xlsx", package="ringdater")
load_chron(chron_path)
}
\keyword{RingdateR}
