% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaded_data_check_function.R
\name{loaded_data_check}
\alias{loaded_data_check}
\title{Check for problems in the loaded data.}
\usage{
loaded_data_check(the_data)
}
\arguments{
\item{the_data}{A dataframe containing the series to be used in RingdateR}
}
\description{
This function checks if there are any problems with data that have been loaded into RingdteR
before running the analysis. It returns an code (0-2) denoting if there are detected problems
with the loaded data.
}
\details{
Output codes:
0 = No problems with the data.
1 = The first column in the dataframe conatins NA values.
2 = The loaded data are discontinuous/contain missing values.
}
\examples{
the_data <- data.frame(x = 1:10, y = 1:10)
loaded_data_check(the_data)
# Return of 0 means no errors
}
\keyword{data}
\keyword{quality}
