% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_series_func.R
\name{plot_all_series}
\alias{plot_all_series}
\title{Plot all data in dataframe along with the arithmetic mean chronology}
\usage{
plot_all_series(aligned_data)
}
\arguments{
\item{aligned_data}{A data frame containing multiple ring width series aligned in time. First column should be years/ring count}
}
\description{
Generate a plot of all the aligned data alogn with the arithmetic mean chronology
}
\keyword{plot}
