% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_bar_EPS_function.R
\name{R_bar_EPS}
\alias{R_bar_EPS}
\title{Calculate the Rbar and EPS of a chronology}
\usage{
R_bar_EPS(the.data, window = 25)
}
\arguments{
\item{the.data}{A data.frame holding individual series in the aligned chronology}

\item{window}{A numeric integer used to calculate the running Rbar and EPS values over.}
}
\description{
This function uses dplR to calculate the Rbar and EPS of a chronology
}
\examples{
chron_path  <- system.file("extdata", "dated_example_excel.xlsx", package="ringdater")
chron_data  <- load_chron(chron_path)
chron_data  <- name_check(chron_data)
chrono      <- normalise(the.data = chron_data, detrending_select = 3, splinewindow = 21)
R_bar_EPS(chrono, window = 21)
}
\keyword{EPS}
\keyword{Rbar}
\keyword{chronology}
