% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_series_function.R
\name{remove_series}
\alias{remove_series}
\title{Remove series that have been dated from the undated dataframe}
\usage{
remove_series(the.data, series.id)
}
\arguments{
\item{the.data}{the dataframe containing the undated series.}

\item{series.id}{A list of sample ID's (column names) that are to be removed from the.data}
}
\description{
This function removes data that have been aligned from the originally loaded undated series in the shiny app.
}
\examples{
undated_path <- system.file("extdata", "undated_example.csv", package="ringdater")
undated_data <- load_undated(undated_path)
samples <- colnames(undated_data )[c(2,3,6)]
remove_series(undated_data, samples)

}
\keyword{data}
\keyword{handling}
